Refactor map model -> style service
