+++ erdblick_app/app/search-menu.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit} from "@angular/core";
+import {Component} from "@angular/core";
 import {Cartesian3} from "cesium";
 import {InfoMessageService} from "./info.service";
 import {JumpTarget, JumpTargetService} from "./jump.service";
